requirements.txt
flask
openai
python-docx
pymupdf
pandas
PyPDF2
